bpf: Look up L4 before L3 and redirect to proxy if either L4 result redirects
Always look up both L3/L4 and L4-only to find out if either of them
redirects to proxy. This allows elimination of additionally generated
policy map entries that would otherwise needed to ensure L4-only
policy redirects to the proxy in the presence of L3/L4 policies that
do not redirect.

This implementation is somewhat constrained by the bpf verifier not
being able to handle two policy lookup results at the same time. This
is why the packet is accounted towards both rules if it matches both
an L3/L4 rule that does not redirect and an L4-only rule. This
double-accounting could be eliminated by repeating the first policy
map lookup, but that would add otherwise unnecessary overhead.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>